refactor orderbook to use Decimal for price representation

+++ src/matching_engine/engine.rs
@@ -1,5 +1,6 @@
-use super::orderbook::{Order, Orderbook, Price};
+use super::orderbook::{Order, Orderbook};
 use std::collections::HashMap;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Eq, PartialEq, Hash, Clone)]
 pub struct TradingPair {
@@ -34,7 +35,7 @@ impl MachiningEngine {
 
     }
 
-    pub fn place_limit_order(&mut self, pair: TradingPair, price: f64, order: Order) -> Result<(), String> {
+    pub fn place_limit_order(&mut self, pair: TradingPair, price: Decimal, order: Order) -> Result<(), String> {
         match self.orderbooks.get_mut(&pair) {
             Some(orderbook) => {
                 orderbook.add_order(price, order);
